fix: broken satellite imagery in GeoMap and GlobalMap (#46)
## Description
Satellite imagery is supposed to be enabled if a google maps api is added to the setting. A button will appear which allows the user to switch between OSM and Google. However that button doesn’t appear now. Google Maps js seems to be loaded correctly so the issue could be in the two components themselves, possible introduced during v3 migration.

## Jira Issue
1. [1327](https://syriajustice.atlassian.net/browse/BYNT-1327)