refactor: change api response for creating/updating a/b/i, fix insertion into table (#75)
**Problem:**
Currently, when creating records on pages such as bulletins, actors,
etc., the newly created item is pushed directly into the table without
an ID. This causes occasional visual artifacts where the table displays
duplicate rows.

**Solution:**
This PR resolves the issue by using the server response, which includes
the correct ID, to update the table. This ensures the integrity of the
data displayed and eliminates duplicate rows caused by the missing ID.

---------

Co-authored-by: cango91 <can.gologlu@gmail.com>