fix: adjust image dims properly after rotation (#146)
When an image file originally saved in landscape orientation is rotated,
it is not fully visible within the display frame. Part of the image is
cropped or hidden after rotation, which affects usability and viewing
experience.

THE FIX IN THIS PR is a workaround since lightgallery.js does not seem
to support adjusting the image to fit after rotation, theres an open
[issue](https://github.com/sachinchoolur/lightGallery/issues/1171) about
it but they never implemented the fix