refactor: use v-autocomplete in SearchField.js (#183)
Refactored SearchField.js component. Visually, it should appear almost the same as the original version. The main change is replacing v-combobox with v-autocomplete.

Previously, the component fetched items from the API as the user typed. When the user pressed enter, it compared the searched value against the list of items to ensure only existing values were inserted. This required a lot of internal validations, increasing the risk of bugs.

By switching to v-autocomplete, we delegate these validations to Vuetify, since the component only allows selecting existing values. This significantly simplifies the component and also addresses the issue of the dropdown not closing.